Fixes and Environment
fixed the camera-to-mouse issue, turns out I forgot to include the line that stops the camera moving when the mouse stops moving. the first room for the environment was mostly finished. also finished the controls for the player and mixed around with the speed and jump strength. still have to figure out how to make the arrow disappear when in game so the arrow doesn't go off screen.